Changed the size of the h1 so it would be larger.
Added css to the info_page div to have it be flexbox. I put the header and the contact information section in a container with each of them as their own item. So I could align them both in the middle.

I also made the contact information it's own container and broke down the different inputs as items. Then used flexbox to align them.